\name{gSSA}
\alias{gSSA}
\title{stochastic simulation function}
\usage{
gSSA(N, R0, gamma, sigma, rho, epsilon, omega, c, m, omegam, mu, mj, phi,
  s, jspan, span, k, preg, Ea0, Ej0, En0, Na, Nj, Nn, whichBeta, whichRho,
  whichGamma, tmax = 10, inc = 52, maxtau = 1/365, nevents = 56)
}
\description{
stochastic simulation function
}
